+++ integrador_2/My Car/mycar/src/main/java/com/example/mycar/dto/RespuestaPagoDTO.java
@@ -7,8 +7,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -16,7 +14,7 @@ import java.math.BigDecimal;
 public class RespuestaPagoDTO {
     private Long facturaId;
     private String numeroFactura;
-    private BigDecimal totalPagado;
+    private Double totalPagado;
     private EstadoFactura estado;
     private TipoPago tipoPago;
     private String mensaje;
